Move gallery to separate file
